Added Text Extractor & Selectable List Attribute
The Text Extractor component is meant to be attached to a Text UI object, and reference another Game Object. It will automatically find all the String Scriptable References attached to the object and display them to the designer, allowing them to select one to be the displayed value.

The Selectable List Attribute is an editor tool that allows the player to select a list of strings and and index integer and it will display all the list in a dropdown menu, updating the index value to match the newly selected item. This can help manage lists in the inspector.